Realigning source.  Moved a ton of placeholder code to the tech.mcprison.prison.spigot package to get it out of the SpigotPlatform class. There was way too much business knowledge being exposed within SpigotPlatform and wanted to group it all under a placeholder's package. The real meat of placeholders has to come together within spigot because core has no knowledge or access to the mines module, or the rank module. This required creating a new interface and pulling the majority of the functions out of the Platform interface.
