+++ prison-core/src/main/java/tech/mcprison/prison/integration/Placeholders.java
@@ -1,0 +1,31 @@
+package tech.mcprison.prison.integration;
+
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
+
+import tech.mcprison.prison.integration.IntegrationManager.PlaceHolderFlags;
+
+public interface Placeholders {
+	
+	public Map<PlaceHolderFlags, Integer> getPlaceholderDetailCounts();
+	
+	
+	public int getPlaceholderCount();
+	
+	
+	public int getPlaceholderRegistrationCount();
+	
+	
+	public String placeholderTranslate(UUID playerUuid, String playerName, String identifier);
+	
+	
+	public String placeholderTranslateText( String text);
+	
+	
+	public String placeholderTranslateText( UUID playerUuid, String playerName, String text);
+	
+	
+	public List<String> placeholderSearch( UUID playerUuid, String playerName, String[] patterns );
+	
+}
